+++ packages/web-ui/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @m5kdev/web-ui
 
+## 0.3.5
+
+### Patch Changes
+
+- organization preferences
+- Updated dependencies
+  - @m5kdev/commons@0.3.5
+  - @m5kdev/frontend@0.3.5
+
+## 0.3.4
+
+### Patch Changes
+
+- deps
+- Updated dependencies
+  - @m5kdev/commons@0.3.4
+  - @m5kdev/frontend@0.3.4
+
 ## 0.3.3
 
 ### Patch Changes